Moving test case to a separate file
